+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to rust_xlsxwriter will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## In progress [0.10.0] - 2022-11-XX
+
+### Added
+
+- Added the [`worksheet.set_freeze_panes()`] method to set "freeze" panes for
+  worksheets. See also the [app_panes] example application.
+
+[app_panes]: https://rustxlsxwriter.github.io/examples/panes.html
+[`worksheet.set_freeze_panes()`]: https://docs.rs/rust_xlsxwriter/latest/rust_xlsxwriter/struct.Worksheet.html#method.set_freeze_panes
+
+
 ## [0.9.0] - 2022-10-31
 
 **Note, this version contains a major backward incompatible API change where it